Move balm bundling to bundling file
