+++ lib/bundling.js
@@ -1,9 +1,13 @@
 const fs = require('fs');
+const traversy = require('traversy');
 const cheerio = require('cheerio');
 const XRegExp = require('xregexp');
 const normalizeSelector = require('normalize-selector');
 const { dirname, basename, resolve, extname, relative } = require('path');
-const { hash, makeScopedClass } = require('./utils');
+const { hash, makeScopedClass, fileExtRegexStr, isDev } = require('./utils');
+const regex = new RegExp(fileExtRegexStr, 'g');
+const { startWatching, notifyClient } = require('./watch');
+const hashes = {};
 
 
 const cleanWhitespace = (str) => str.replace(/\s+/g, ' ').trim();
@@ -263,7 +267,68 @@ function extractBlocks(path, root, blocksDir) {
 const fixInlineAttributeConditions = (str) => str.replace(/=\"=(=?)\"/gm, '==$1');
 
 
-exports.blockTypes = blockTypes;
-exports.extractBlocks = extractBlocks;
-exports.addHmrToBundles = addHmrToBundles;
-exports.writeEntryFile = writeEntryFile;
+function bundle(opts) {
+  const { blocksDir, root, port, watch } = opts;
+
+  blockTypes.css.bundles = { entry: [] };
+  blockTypes.js.bundles = { entry: [] };
+
+  if (!fs.existsSync(blocksDir)) {
+    fs.mkdirSync(blocksDir);
+  }
+
+  traversy(root, fileExtRegexStr, (path) => {
+    bundleFile(path, opts);
+  });
+
+  if (isDev) {
+    addHmrToBundles(port, blocksDir);
+  }
+
+  if (watch) {
+    startWatching(opts);
+
+    fs.watch(root, { recursive: true }, (_, path) => {
+      if (extname(path) === '.balm') {
+        bundleFile(resolve(root, path), opts);
+      }
+    });
+  }
+
+  writeEntryFile('css', blocksDir);
+  writeEntryFile('js', blocksDir);
+}
+
+
+function bundleFile(path, opts) {
+  const { root, blocksDir, loadHandles, handles, watch, templates } = opts;
+  const key = path.replace(regex, '');
+  const blocks = extractBlocks(path, root, blocksDir);
+
+  Object.entries(blocks).forEach(([ _, block ]) => {
+    const { type, file, content, contentHash } = block;
+    const path = `${blocksDir}/${file}`;
+    const previousHash = hashes[file];
+
+    if (contentHash !== previousHash) {
+      fs.writeFileSync(path, content);
+      hashes[file] = contentHash;
+
+      if (loadHandles && type === 'handle') {
+        delete require.cache[require.resolve(path)];
+        handles[key] = require(path);
+      }
+
+      if (watch && previousHash && ['template', 'handle'].includes(type)) {
+        notifyClient(path);
+      }
+
+      if (type === 'template') {
+        templates[key] = content;
+      }
+    }
+  });
+}
+
+
+exports.bundle = bundle;
